Factor out common GUI code and decouple GUI controls from windows.
The new framework exposes the code to create a GUI for any given
generator, and to access the generator controlled by the code.
That code is no longer wrapped up in the EditorWindow subclass, a
separate class which owns an instance of the object containing the
GUI code now creates the Editor Window.

This will allow us to write GUIs that combine various generators,
such as filter debuggers which offer a choice between generators.