Create a registration/lookup system for components.
Components (generators and filters) and their associated GUIs are
registered using an attribute, one of Generator, Filter,
GeneratorGUI, or FilterGUI. These attributes associate the
implementation with a string, with which the component can be
queried via the language framework. In the case of the Generator
and Filter attributes, the implementer can also specify the required
and optional arguments for instantiating component, along with the
types of the arguments and default values for optional arguments.
The language framework uses this information to generate runtime
typechecking code for each generator and filter.